backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jaechan <dlwocks31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
pymysql = "^1.1.0"
python-decouple = "^3.8"
gunicorn = "^21.2.0"
boto3 = "^1.28.62"
drf-yasg = "^1.21.7"
pillow = "^10.0.1"
requests = "^2.31.0"
coverage = "^7.3.2"
sentry-sdk = {extras = ["django"], version = "^1.32.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
